+++ src/main/resources/shaders/interface/simple.vert
@@ -1,6 +1,6 @@
 #version 430
 
-layout (location = 0) in float vertexId;
+layout (location = 0) in int vertexId;
 
 layout(std430, binding=0) buffer transforms {
     mat4 matrices[];
@@ -9,6 +9,6 @@ layout(std430, binding=0) buffer transforms {
 out mat4 pass_transform;
 
 void main() {
-    pass_transform = mat4(matrices[int(vertexId)]);
+    pass_transform = mat4(matrices[vertexId]);
     gl_Position = vec4(0, 0, 0, 1);
 }
